Major refactoring to improve modularity and reenabled unit tests.
Created examples folder.

BVH acceleration structure implemented.

Orientation.h -> Transform.h

Instance.h: Instancing is now implemented. Rotations and Translations.

Material.h: changed default Lambertian material color to be more noticeable.

ProgressBar.h: implemented a reusable progress bar and uses \r to avoid writing too many lines to the console.

README.md: updated to reflect changes to API.

Sphere.h -> ImplicitShape.h

Ray.h: Ray directions are always unit vectors.

Utility.h: abs(), sign(), mix(), random_int()

Vec3.h: New constructor Vec3(double s), rotate(), operator-(), abs(), max(), min(), max_element(), and min_element().

Camera.h: added progress bar and changed class Orientation -> Transform.

+++ src/Image.cpp
@@ -165,7 +165,7 @@ Image Image::to_ascii_image()
             old_color /= ASCII_LENGTH;
 
             // Use the intensity to determine which character should be used for the 
-            const int brightness = (old_color.x() + old_color.y() + old_color.z())*(ASCII_TABLE_LENGTH-1)/3.0;
+            const int brightness = static_cast<int>((old_color.x() + old_color.y() + old_color.z())*(ASCII_TABLE_LENGTH-1)/3.0);
             const bool *character = ASCII_TABLE[brightness];
             
             // This loop can be executed in parallel and enables resuse of the above data.
